for about page
text: 

Give. Make.

Every day, people around the world send millions
of tons of waste to landfills. Why? 

Often the answer is that it’s cheaper and easierthan any other option. But it doesn’t have to be. 

Chaff map is here to provide place to share
waste so that it may be used for an alternate 
purpose and stay out of our landfills.

We have one house, let’s keep it clean. 

Sign up.Give. Make.